move synchronisation code to own class, set new startup activity with does syncrhonisation and setup list, and starts after it the Listactivity
