+++ app/src/main/java/com/sw_ss16/studyroompopulationpredicter/Synchronisation/Sync.java
@@ -1,0 +1,148 @@
+package com.sw_ss16.studyroompopulationpredicter.Synchronisation;
+
+import com.android.volley.Request;
+import com.android.volley.RequestQueue;
+import com.android.volley.Response;
+import com.android.volley.VolleyError;
+import com.android.volley.toolbox.JsonArrayRequest;
+import com.sw_ss16.studyroompopulationpredicter.backend.Database;
+
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
+
+/**
+ * Created by daniel on 07.04.16.
+ */
+public class Sync {
+
+    public static void insertStudyRoomsIntoSQLiteDB(RequestQueue queue, final Database db) {
+        String url = "http://danielgpoint.at/predict.php?what=lc&how_much=all";
+
+        final JsonArrayRequest jsonArrayRequest = new JsonArrayRequest(Request.Method.GET, url , null,
+                new Response.Listener<JSONArray>() {
+                    @Override
+                    public void onResponse(JSONArray response) {
+                        for (int i = 0; i < response.length(); i++) {
+                            try {
+                                JSONObject jsonObject = response.getJSONObject(i);
+                                String id = jsonObject.getString("id");
+                                String name = jsonObject.getString("name");
+                                String description = jsonObject.getString("description");
+                                String address = jsonObject.getString("address");
+                                String capacity = jsonObject.getString("capacity");
+                                //TODO: Image IN
+                                //TODO: Image OUT
+                                System.out.println(id + " " + name + " " + address);
+                                db.insertInDatabase("INSERT INTO studyrooms (ID, NAME, DESCRIPTION, ADDRESS, IMAGE_IN, IMAGE_OUT, CAPACITY) " +
+                                        "SELECT " +
+                                        id + "," +
+                                        "'" + name + "'," +
+                                        "'" + description + "'," +
+                                        "'" + address + "'," +
+                                        "null," + // TODO: image
+                                        "null," + // TODO: image
+                                        capacity + " " +
+                                        "WHERE NOT EXISTS (SELECT 1 FROM studyrooms WHERE ID = " + id +");");
+
+                            } catch (JSONException e) {
+                                e.printStackTrace();
+                            }
+                        }
+                    }
+                }, new Response.ErrorListener() {
+            @Override
+            public void onErrorResponse(VolleyError error) {
+                System.out.println("That didn't work!");
+            }
+        });
+        // Add the request to the RequestQueue.
+        queue.add(jsonArrayRequest);
+    }
+
+    public static void insertStatisticsIntoSQLiteDB(RequestQueue queue, final Database db) {
+        String url = "http://danielgpoint.at/predict.php?what=stat&how_much=all";
+
+        final JsonArrayRequest jsonArrayRequest = new JsonArrayRequest(Request.Method.GET, url , null,
+                new Response.Listener<JSONArray>() {
+                    @Override
+                    public void onResponse(JSONArray response) {
+                        for (int i = 0; i < response.length(); i++) {
+                            try {
+                                JSONObject jsonObject = response.getJSONObject(i);
+                                String id = jsonObject.getString("id");
+                                String lc_id = jsonObject.getString("lc_id");
+                                String weekday = jsonObject.getString("weekday");
+                                String hour = jsonObject.getString("hour");
+                                String fullness = jsonObject.getString("fullness");
+                                System.out.println(id + " " + lc_id + " " + weekday);
+                                db.insertInDatabase("INSERT INTO statistics (ID, LC_ID, WEEKDAY, HOUR, FULLNESS ) " +
+                                        "SELECT " +
+                                        id + "," +
+                                        "" + lc_id + ", " +
+                                        "" + weekday + ", " +
+                                        "" + hour + ", " +
+                                        "" + fullness + " " +
+                                        "WHERE NOT EXISTS (SELECT 1 FROM statistics WHERE ID = " + id +");");
+
+                            } catch (JSONException e) {
+                                e.printStackTrace();
+                            }
+                        }
+                    }
+                }, new Response.ErrorListener() {
+            @Override
+            public void onErrorResponse(VolleyError error) {
+                System.out.println("That didn't work!");
+            }
+        });
+        // Add the request to the RequestQueue.
+        queue.add(jsonArrayRequest);
+
+    }
+
+    public static void insertCurrentDataIntoSQLiteDB(RequestQueue queue, final Database db) {
+        String url = "http://danielgpoint.at/predict.php?what=curr&how_much=all";
+
+        final JsonArrayRequest jsonArrayRequest = new JsonArrayRequest(Request.Method.GET, url , null,
+                new Response.Listener<JSONArray>() {
+                    @Override
+                    public void onResponse(JSONArray response) {
+                        for (int i = 0; i < response.length(); i++) {
+                            try {
+                                JSONObject jsonObject = response.getJSONObject(i);
+                                String id = jsonObject.getString("id");
+                                String lc_id = jsonObject.getString("lc_id");
+                                String date = jsonObject.getString("date");
+                                String hour = jsonObject.getString("hour");
+                                String fullness = jsonObject.getString("fullness");
+                                System.out.println(id + " " + lc_id + " " + date);
+                                db.insertInDatabase("INSERT INTO current_data (ID, LC_ID, HOUR, FULLNESS, DATE) " +
+                                        "SELECT " +
+                                        id + "," +
+                                        "" + lc_id + ", " +
+                                        "" + hour + ", " +
+                                        "" + fullness + ", " +
+                                        "'" + date + "' " +
+                                        "WHERE NOT EXISTS (SELECT 1 FROM current_data WHERE ID = " + id +");");
+
+                            } catch (JSONException e) {
+                                e.printStackTrace();
+                            }
+                        }
+                    }
+                }, new Response.ErrorListener() {
+            @Override
+            public void onErrorResponse(VolleyError error) {
+                System.out.println("That didn't work!");
+            }
+        });
+        // Add the request to the RequestQueue.
+        queue.add(jsonArrayRequest);
+
+    }
+
+    public static void insertFavoriteStudyRoomsIntoSQLiteDB(Database db) {
+
+    }
+}
